Rename / revamp vebal remapper deployment task.
