--- pkg/deployments/tasks/20230403-vebal-remapper/readme.md
@@ -1,8 +1,0 @@
-# 2023-04-03 - Voting Escrow Remapper
-
-Deployment of the `VotingEscrowRemapper` contract, which allows delegating veBAL balances in mainnet to a different address in L2s and sidechains.
-This feature is useful for contract accounts with veBAL in mainnet, since the same address might not be accessible in other networks.
-
-## Useful Files
-
-- [`VotingEscrowRemapper` artifact](./artifact/VotingEscrowRemapper.json)
